+++ Mesh/src/notes.txt
@@ -1,3 +1,14 @@
 Flash Memory
 Boolean: powerCheck
 Integer: defaultPower
+
+/*
+    Plan:
+    - Create sync function to sync each node to a particular time
+    - Measure data 4x per day, and send to gateway
+    - For pulse: measure time between pulses
+    - Determine power to cover distance
+        - Rerun power analysis
+        - Do we want dynamic power to radio?
+
+*/
